End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -4,6 +4,7 @@ import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
 import { TimeagoIntl } from 'ngx-timeago';
 import { strings as englishStrings } from 'ngx-timeago/language-strings/hr';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -18,7 +19,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService, intl: TimeagoIntl) {
+  constructor(private messageService: MessageService, intl: TimeagoIntl, public confirmService: ConfirmService) {
     intl.strings = englishStrings;
     intl.changes.next();
   }
@@ -46,9 +47,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id),1);
+    this.confirmService.confirm('Confirm delete message','This cannot be undone.').subscribe(result => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id),1);
+        })
+      }
     })
+   
   }
 
 }
